+++ src/internal/utils/date-time/format-date-range.ts
@@ -1,7 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { formatTimezoneOffset, isIsoDateOnly } from '.';
+import { isIsoDateOnly } from './is-iso-date-only';
+import formatDateIso from './format-date-iso';
+import formatDateLocalized from './format-date-localized';
 
 export type AbsoluteDateFormat = 'iso' | 'long-localized';
 
@@ -43,7 +45,7 @@ export function formatDateRange({
 }
 
 function formatDate({
-  date: isoDate,
+  date,
   format,
   hideTimeOffset,
   isDateOnly,
@@ -59,81 +61,10 @@ function formatDate({
 }) {
   switch (format) {
     case 'long-localized': {
-      const date = new Date(isoDate);
-
-      const formattedDate = new Intl.DateTimeFormat(locale, {
-        day: 'numeric',
-        month: 'long',
-        year: 'numeric',
-      }).format(date);
-
-      if (isDateOnly) {
-        return formattedDate;
-      }
-
-      const formattedTime = new Intl.DateTimeFormat(locale, {
-        hour: '2-digit',
-        hourCycle: 'h23',
-        minute: '2-digit',
-        second: '2-digit',
-      }).format(date);
-
-      const isRTL = locale && getDirection(locale) === 'rtl';
-
-      const formattedDateTime = isRTL
-        ? [formattedTime, formattedDate].join(getDateTimeSeparator(locale)) + '\u200E' // Add LTR mark at the end to be able to concatenate correctly to form the date range.
-        : [formattedDate, formattedTime].join(getDateTimeSeparator(locale));
-
-      if (hideTimeOffset) {
-        return formattedDateTime;
-      }
-
-      const formattedTimeOffset = formatTimezoneOffsetAbsolute(isoDate, timeOffset);
-      if (isRTL) {
-        return [formattedTimeOffset, formattedDateTime].join(' ');
-      }
-      return [formattedDateTime, formattedTimeOffset].join(' ');
+      return formatDateLocalized({ date, hideTimeOffset, isDateOnly, locale, timeOffset });
     }
-
     default: {
-      const formattedOffset = hideTimeOffset || isDateOnly ? '' : formatTimezoneOffset(isoDate, timeOffset);
-      return isoDate + formattedOffset;
+      return formatDateIso({ date, hideTimeOffset, isDateOnly, timeOffset });
     }
   }
-}
-
-// Languages written from right to left (RTL)
-const rtlLanguages = ['ar', 'he'];
-
-// Languages in which date and time are separated just with a space, without comma
-const languagesWithoutDateTimeSeparator = ['ja', 'zh-CN'];
-
-function getDirection(locale: string) {
-  return rtlLanguages.includes(getPrimarySubTag(locale)) ? 'rtl' : 'ltr';
-}
-
-function getDateTimeSeparator(locale?: string) {
-  if (!locale) {
-    return ', ';
-  }
-  return languagesWithoutDateTimeSeparator.includes(locale) ? ' ' : getDirection(locale) === 'rtl' ? ' ,' : ', ';
-}
-
-function getPrimarySubTag(locale: string) {
-  return locale.split('-')[0];
-}
-
-function formatTimezoneOffsetAbsolute(isoDate: string, offsetInMinutes?: number) {
-  offsetInMinutes = offsetInMinutes ?? 0 - new Date(isoDate).getTimezoneOffset();
-  if (offsetInMinutes === 0) {
-    return '(UTC)';
-  }
-  const hoursOffset = Math.floor(Math.abs(offsetInMinutes) / 60);
-  const minuteOffset = Math.floor(Math.abs(offsetInMinutes % 60));
-
-  const sign = offsetInMinutes < 0 ? '-' : '+';
-  const formattedOffsetMinutes = minuteOffset === 0 ? '' : `:${minuteOffset}`;
-  const formattedOffset = `(UTC${sign}${hoursOffset}${formattedOffsetMinutes})`;
-
-  return formattedOffset;
 }
