+++ src/internal/utils/date-time/index.ts
@@ -6,7 +6,7 @@ export { formatDateRange, AbsoluteDateFormat } from './format-date-range';
 export { formatDate } from './format-date';
 export { formatTime } from './format-time';
 export { formatDateTime } from './format-date-time';
-export { formatTimezoneOffset } from './format-timezone-offset';
+export { formatISOTimezoneOffset } from './format-timezone-offset';
 export { isIsoDateOnly } from './is-iso-date-only';
 export { joinDateTime, splitDateTime } from './join-date-time';
 export { parseDate } from './parse-date';
